Make A/B test meta tag consistent with Google Analytics JS
Reformat the meta tag which specifies user's bucket for the A/B example
test so that it matches the format expected by the analytics JavaScript
code in `static`.

The previous format was only parsed by `govuk-toolkit-chrome`, but that
code has not been released yet so it's safe to change this tag and then
update the Chrome plugin.