Allow app to determine schema of requested content from content-store
Adds a ContentFormatInspector class which is responsible for finding the
content format for a requested slug, and handling API errors. This is
used by the FormatRoutingConstraint in the router to help determine which
controller to route the request to.

Since we're now in the middle of migrating mainstream content to
content-store there is a fair amount of complexity being introduced. Since
content might need to be pulled from either the content-api or the content-
store, the Frontend app needs to be able to decide which to call. It's hoped
that it will be relatively easy to reduce this complexity once all of the
mainstream content has been migrated.