Fall back to English when content is in an unsupported locale
The content schema locales include some locales which are not supported
by Rails by default, such as `tk` and `zh`. This triggers an exception
in ApplicationController when it attempts to set the application locale
from the content locale.

This was causing flaky tests when a random content item was generated.
For example, help_controller_test would fail about 20% of the time when
a content item with an unsupported language was generated. Example seed:
47089.

This fix checks whether the content's locale is in the list locales
supported by the Rails application, and falls back to English if not.

A better long-term fix might be to align the locales in
govuk-content-schemas with the available locales in GOVUK Rails
applications.