Add questions to assisted digital help with fees survey
The extended version of this form is based on feedback from the team that
requested it in the first place.

The form is now much longer and so requires interactive showing and hiding
of elements. We get this from GOVUK.ShowHideContent() from
govuk_frontend_toolkit which we initialise only for this page, if more
pages need similar functionality we can add it to the main JS bundle for
the whole of frontend.