Set token in default url options for fact check content
For content that is in fact check a JWT token is generated and the
authenication proxy will pass that through with the request to the
frontend.

This token needs to be carried through on the get and post requests for
those formats that are a little more interactive.

The authentication proxy no longer strips the token itself as a result
of [this PR](https://github.com/alphagov/authenticating-proxy/pull/21)

https://trello.com/c/RoHFOwid/646-enable-postcode-search-for-local-transactions-in-factcheck