Render simple smart answer format
Update controller to read content from content store, add specialised presenter for the schema. We remove any tests to do with draft, as we don't use editions anymore to determine if a draft item is being viewed. Also there is no cacheing happening on the draft stack so we can remove the code that conditionally set the expiry headers. We also remove the fixtures for 'bridge of death' and prefer to move it into the test itself.