Create GuidePresenter from PublicationWithPartsPresenter
It was noted in the review feedback for this Pull Request that
the emergent naming convention for presenters is by format, so this
introduces a GuidePresenter which wraps the access methods for Parts.

In doing this, I encountered some missing tests for the original presenter
and this led me a little way down the rabbithole.