Get licence data from Licensify
Previously licence information was added to a content item (artefact) in the ContentAPI (via a call to Licencify). This commit updates Frontend to make the call to Licensify instead, copying the pattern that local transactions and places follow.

Also updated the licence test as it was not stubbing calls to Licencify correctly. The first call for a location_specific licence should not be returning any authorities. Only when a call to licensify with a snac code is made should it include authorities.