Don't hardcode static.dev.gov.uk in slimmer initializer
In development we can use the env var 'STATIC_DEV' to tell frontend to
configure slimmer to use a different asset host for static.  If you don't
set that var it falls back to the default static host on dev which is
currently hardcoded to be `http://static.dev.gov.uk`.

However, this doesn't use Plek so we can't take advantage of the existing
env vars that Plek uses to allow customising service uris (for example:
`PLEK_SERVICE_STATIC_URI`).  The default in Plek for asset_host if not
configued is to use `Plek.current.find('static')` which is what we want as
it will take advantage of those env vars.  If we remove all the
`config.slimmer.asset_host` references we get the benefit of the default
behaviour and can use the existing env vars to control which host to use
for static.

The other advantage of using the existing env vars is that they work across
all places that we might need static, (e.g. the host that slimmer uses to
fetch templates for things like govuk-components, see:
https://github.com/alphagov/slimmer/blob/master/lib/slimmer/component_resolver.rb#L54
) as well as just configuring the asset_host.

[Fixes #1145]