Rename Licence 'search' controller method to 'start'
This feels more appropriate as we only show the postcode search form for about ~20 of all licence pages. The rest of the time we either present a link to an external website or redirect to the single licensing authority. The code also supports a radio button for multiple licensing authorities but there is currently no example if this in Licensify.