Link from the homepage to the Education taxon in the B variant
In the new navigation, which is the B variant of the education
navigation A/B test, link directly to the Education taxon rather than
the Education browse page, and use the title and description of the
taxon.

The layout of the homepage has been changed slightly to accommodate the
longer taxon description: 'Housing and local services' has been moved
from the second column into the third. This matches the prototype that
has been tested in user research.

The variants are cached separately using the `Vary` header, but unlike
on other A/B-tested pages, the `meta` tag is not added. This is because
we do not want to include *all* journeys through the homepage in the A/B
test analytics; we are only concerned with user journeys which include
education content.