Show old related links for some mainstream pages
Until we have the ability to create curated related links for the new
taxonomy, we need to still show the old sidebar because the links shown
in there are quite important.

This is a temporary fix that allows us to show the old sidebar on the
new navigation for a number of mainstream pages.

When `content-tagger` allows to do this, we should revert this commit.