Use banner ID specific to the taxonomy survey
Like the banner-notification it replaces, this will ensure that the
standard site survey isn't displayed when we're showing the taxonomy
one. Using a custom ID means we don't have to try and reverse styling
applied to banner-notification. The existing use of display:none doesn't
actually work as expected because survey.js in static looks for a
visible element when deciding whether or not to show the site survey.