Fix location specific licences with multiple authorities
We [recently refactored Licences](https://github.com/alphagov/frontend/pull/1141) and missed this edge case where a location specific licence search (eg for a postcode in Greenwich) could return multiple licence authorities. We found this for the licence "street-collection-licence". The code previously would return the first of the authorities, so we do the same here. In the case of this licence it doesn't seem like the right thing to be doing as the start pages advises to contact the Metropolitan Police for London (which is the second authority which would not get rendered), but at this point we're just concerned with restoring the original functionality of the code.