Move banner into partial
This ensures that we can deploy most of the functionality in advance and switch over the banner content in the final deploy.

As we want to be able to deploy just the content, the tests need to support
a partial both with and without content. To do this, we override the view path to stub templates
replacement. This is borrowed from rspec-rails’ `stub_template` method
https://github.com/rspec/rspec-rails/blob/e8054a1cd03044f725030fe8315952
cf3799a395/lib/rspec/rails/example/view_example_group.rb#L86-L88