Add analytics meta tags to homepage and search page
Load the content item for the homepage and search page so that it is
used to populate the analytics meta tags.

This will enable us to track these pages as `finding` rather than
`thing` in the navigation analytics. Currently they make the user
journey analytics rather misleading because the appear to be content
pages.

Delete the old slimmer configuration, since it will be replaced by the
meta tag from govuk-components.