No longer selectively set the Cache-Control header
Now that Frontend is used as a draft app reading content from the
draft content store, the caching configuration is no longer something
it should manage.

Previously the existence of the `edition` parameter was used to detect
if a draft was being requested, but this no longer works, as it is not
possible for Frontend to request a specific edition from the
relevant (draft or live) Content Store service.