Displaying promo depend on emergency banner
The promo banner must only display if no emergency banner is enabled. We
are querying Redis to discover this. This commit ensures that the promo
banner will only display if the emergency banner is not set in Redis.

* Pass an extra boolean variable to the `index.html.erb` from
`homepage_controller` to indicate the state of the emergency banner
(enabled or disabled).

* Add a condition in `index.html.erb` to selectively display the promo
banner based on this new local variable.

* Amend a couple of tests for the promo banner.