Add “register to vote” promo section
In the run-up to the snap election, we have agreed with No. 10 to
replace the GOV.UK blogs feature promo slot to promote the register to
vote service start page.

The GOV.UK blogs feature will be replaced after the registration
deadline for the snap election passes, and we should then revert this
commit.