Add SRI attributes to javascript and stylesheets
We rely on sprockets-rails to do the heavy lifting for us here and
calculate the integrity sha for these assets.  We also add th ecrossorigin
attribute set to 'anonymous' because our assets are served from a
different domain to the main application (assets.publishing.service.gov.uk
vs www.gov.uk in most cases) and so to avoid dealing with CORS we tell the
browser to not send CORS headers with the request for the asset.