Apply simple_format to access_notes and general_notes in places
Some imminence users requested more formatting options for notes and
using the rails `simple_format` is an easy first attempt at allowing
this.  There's precedent for doing this already as we do the same for
some details of licence and travel advice API details.  Adding
something like govspeak would be possible but we'd have to do the
rendering in imminence as no frontend apps handle rendering govspeak
themselves, and it would require changing the schema.