Show error messages for empty submissions on simple smart answers
We follow the pattern than rails checkbox helper uses to ensure we always
have a response param when submitting the form by adding a hidden field
tag with an empty value.  When given two params with the same name Rails
will choose the last one as the value reported when asking for a param
in the controller, so when the user makes a choice we get two response
params; the blank one from the hidden field, and the choice they made
from the radio buttons.  The radio button one will be last and so it's
the value that Rails will report for the response param.

This means we can tell the difference between someone visiting the
smart-answer part way through with a direct url and someone on the
previous step pressing the "Next step" button.  A direct URL visit won't
have a response param, but the "Next step" button submission will, so we
can insert an error message about answering the question.