Add transaction pages to list of mainstream links
Recently, some transaction pages have been moved from smart-answers
into this application. Two of those pages had overrides to show the old
related links instead of the new taxonomy related links when presented
in the B variant of the Education Navigation A/B test.

This commit adds those 2 pages to the static list of pages with
overrides, and adds an integration test to explain why that list exists.

This was picked up by Smokey tests.

There is more context in:
- https://trello.com/c/J8ofEnrt/167-display-curated-related-links-in-the-sidebar
- https://docs.google.com/document/d/1k1nlQOpFwjaUNoNtizAnsDr5z9O5uUMIOlgDSUCbwj4/edit#