Bump govuk_schemas version
When doing schema testing CI is somehow failing with this message:

```
Downloading govuk_schemas-2.1.0 revealed dependencies not in the API or
the
lockfile (json-schema (~> 2.5.0)).
Either installing with `--full-index` or running `bundle update
govuk_schemas`
should fix the problem.
```

Somehow the gem seems to have been installed without taking in the
dependency.