Track simple smart answer completion in GA
In the smart-answers app we track when the user reaches the final outcome
stage of the smart answer by sending a message to GA.  This commit adds
the same functionality to simple smart answers.

We do it via a JS module for tracking smart answers.  Currently it only
tracks the outcome node type and sends a "Smart Answer", "Completed" event
to GA with the path and smart answer name, but there is scope to extend
the module to track other node types and events if we need to.  Note that
the standalone smart answers app only tracks completion at the moment too
so any extra tracking would need to happen there too.