Content update to benefits section
According to the content design team 'Universal Credit’ was the most
searched for term from the GOV.UK homepage from 1 July 2016 30 June
2017.

In the 1st half of that year, there were 105,887 searches. In the last
6 months, there were 218,385 searches.

This number is on trend to grow, especially as Universal Credit is
rolling out  to more people who are on existing benefits and the UC
availability is becoming available in more areas to new claimants.

Our stats show that users aren’t finding their way through Benefits
browse categories or related links on pages. We’ve added links to UC in
Browse, but plans to curate and audit the browse pages have been put on
hold for the time-being.

Hence the reason behind this content update.