Add BOM remover gem to fix SRI issues on firefox
This injects itself into the rails asset pipeline and makes sure that
the CSS we compress with sass does not include a BOM if it is a UTF-8
file.  Firefox < 52 has a bug in how it calculates SRI hashes for CSS
files with a BOM and this gem mitigates that.  Currently none of the
delivered CSS assets are UTF-8, so it wouldn't be a huge problem if
we didn't use this, but it hasn't always been the case and we don't
want the addition of one character in a CSS file to break things in
the future for some users.