Short circuit "ENTERPOSTCODE" requests
We're currently getting around a million requests per week for this postcode.
The requests come from the same IP address and we're pursuing a P&E resolution
to this.  In the meantime, short circuit this one so we don't spend time on API
responses when we know the result.

The source IP does send reasonable requests as well and appears to be affiliated
with local gov, so we don't want to ban it at this point.

The requests are only coming in to the /get-on-electoral-register URL so we
don't need to worry about licence/find local council or other postcode lookup
end points at the moment.