Add LGD to the area types that we support
We support Northern Ireland councils now, but we need to allow them to be found.
This area type is already supported in [find-local-council](https://github.com/alphagov/frontend/blob/master/app/controllers/find_local_council_controller.rb#L8)
and by adding it to the list here, will be activated in local transactions as
well.

We should have done this a while ago; it seems to have been an oversight.