Remove custom helper for content item in find local council
In 490d01f1bdc46be64a0e8ba6ef67ccb81a94dafb we replaced the hardcoded
content item response in the `FindLocalCouncilController` with a version
that spoke to the content store.  However, we also inlined our own
`setup_content_item_and_navigation_helpers` method instead of re-using
the version in `ApplicationController`.  This means we stopped setting
the `govuk:section` meta tag, but also means we'll miss out on any
other changes that we want to apply globally.  We revert back to using
the `ApplicationController` version to avoid this.