Add server side GA tracking for postcode errors
For: https://trello.com/c/6qGrfNOF/238-server-side-tracking-further-discussion

This is part of a spike to investigate whether it is possible to send certain GA events via server side code in order to avoid adding extra load on frontend resources for tracking.
Our currentl solution uses GA tracking via JS, which increases the load time for a page.

With this alternative version, we can leave it up to the server to send along this information, thus avoiding writing extra JS code.