Add configuration for tasklist sidebar
This is the content to be used in the learn-to-drive-a-car tasklist sidebar.

We also add some tests to help ensure that the file can be parsed and is in the
sort of structure that the tasklist component will expect.

The tasklist component is currently in review in Static, so we're reliant on
the data structure for that not changing.