Show tasklist on TaskListSidebar b variants
This is for the tasklist public beta.  The first theme is learning to drive a
car.  This commit adds a tasklist sidebar to content served by Frontend that
is listed in app/controllers/concerns/tasklist_ab_testable.rb.

The content in Frontend that is relevant is limited to transactions and
simple_smart_answers at present.

The tasklist is configured using the file at config/tasklists/learn-to-drive-a-car.json

Testing of the component itself is limited as testing components from static is
not currently possible (as noted in https://github.com/alphagov/slimmer#testing-components)

This limitation is subject to an RFC at present.  What we can do is test the
json that Slimmer emits in the test - which is what we've done.

We need to get a design review on this before merging.