Add "Sign in" to locale files
Publishers are now able to select “Sign in” as start button text, however the
`cy` locale does not contain a Welsh translation, meaning that, incorrectly,
the English text would be displayed.  This isn’t correct, so we update the `cy`
locale with the Welsh translation "Mewngofnodi”, and add “Sign in” to the `en`
locale so we can make use of I18n to retrieve the correct text for a piece of
content’s locale.