Enable Welsh "Start now" translation to be available
A bug has occurred where Welsh content displays “Start now” as start button
text instead of the Welsh translation “Dechrau nawr”.  Previously the
`start_button_text` value on the content item was used as the button text, and
only if this is empty would the locale value be used.  Since the
`start_button_text` value is populated by Publisher and will always be English,
this meant that the Welsh translation would never be used.  This commit changes
the presenter logic so that the presence of a locale value is checked first,
and the `start_button_text` value is only used if no locale value is available.
We also ensure that if there is no `start_button_text` value available then the
default “Start now” locale value is used.