Remove open in new window feature for transactions
Proposal to remove this so we don't need to add it as a feature when consuming the [button component](https://govuk-static.herokuapp.com/component-guide/button/start_now_button_with_info_text)

It's not clear why this was added, but it's wildly inconsistent since we don't seem to be updating it.

Be good to get any input in on if this is a good idea to remove.

Originally added in https://github.com/alphagov/frontend/pull/43