Upgrade to Rails 5.1.4
Main changes required revolved around the changes to how params are
passed within tests when calling get. Rather than passing key/values
when calling get, you now need to pass them within a params hash.

In addition to this, the params hash within controllers is now an
object which has to explictly allow parameters.

Finally, we upgraded anything that was required to be upgraded to be
compatible with Rails 5