Move heading into link with image, remove duplicate link in copy
This has been a long standing accessibility issue that we duplicate content in our promo slots.

This change reduces the amount of time we're linking to one, and keeps the thumbnail clickable.

I've intentionally left the heading looking like a link to make it obvious to all users that they can click it.

This also allows us to remove the duplicate link from the copy below.