[WIP] Remove CSRF Token validation
When migrating to Rails 5 - posting forms was broken to cause all post
requests to be unable to validate the authenticity token, despite
the token being passed through with the params correctly.

As the application has no sense of users/sessions, it makes no sense
to use CSRF protection, and we have removed it.