Moves Tasklist functionality to GovukNavigationHelpers
We are moving the Tasklist functionality into GovukNavigationHelpers so
we can consolidate the code in one place.

Moving shared logic to the gem has changed the way we need to test, so we'll
test real things that we expect to be in the task list.

In addition, we can now test the real component because it is now contained
in a gem locally, so we can assert on actual rendered classes rather than the
json generated by the test helpers for static.