When setting LOWEST_TIER_AREA_TYPES don't mutate UNITARY_AREA_TYPES
Freezing the UNITARY_AREA_TYPES array exposes the fact that when we set
LOWEST_TIER_AREA_TYPES we were also mutating UNITARY_AREA_TYPES because
ultimately the line:

    LOWEST_TIER_AREA_TYPES = UNITARY_AREA_TYPES << DISTRICT_AREA_TYPE

is parsed as

    LOWEST_TIER_AREA_TYPES = ( UNITARY_AREA_TYPES << DISTRICT_AREA_TYPE )

and `<<` is a mutation method on an array, not one of the methods that
returns a new array.

Ultimately, this shouldn't have affected any output in this controller
because UNITARY_AREA_TYPES is only used in this line where we use it to
set LOWEST_TIER_AREA_TYPES so we've been lucky.  If it was used we'd run
the risk of presenting info about a district council when we only wanted
to present info about Unitary councils.