Fix outstanding Date & Time lint offenses
Mostly this is us swapping from `DateTime` to `Time.zone` which makes
sure that we don't forget about timezones when we parse and render date
and time information.

This does mean we've had to introduce a change to the time zone
configuration of the application; it was using the default time zone of
UTC, but we've now configured it to use London instead (like whitehall
and government-frontend).  We spotted this when looking at the travel
advice atom feed.  The rails atom builder will call `xmlschema` on the
object used to generate the updated element and the behaviour of a
`DateTime` object and a `Time` object is different.  A `DateTime` object
will output `+00:00` when the timezone is UTC, but a `Time` object will
output `Z` (the special case for UTC allowed by [the ISO 8601
spec][iso8601-spec]).  By changing the application time zone to 'London'
a `Time` object will now output `+00:00` instead of `Z` (or `+01:00` in
summer when we're in BST).

We also have to define an explicit `feed_id` method on the
`TravelAdviceIndexPresenter::IndexCountry` objects as the behaviour of
`to_s` for a `DateTime` and a `Time` object is different; `DateTime`
returns an ISO8601 formatted string, but `Time` does not. We don't want
the ids of these entries to change, so we've made a method to generate the
ids from a `Time` object that matches the behaviour we would have got from
calling `to_s` on a `DateTime` object.

[iso8601-spec]: https://en.wikipedia.org/wiki/ISO_8601#Time_zone_designators)