Remove wait_until and with/without_javascript helpers
Capybara does an execellent job of making sure you don't need to use
something like wait_until so we shouldn't need this (indeed we don't use
it directly in any tests).  We do use it in our `assert_current_url`
helper, but the code inside wait_until only runs for the javascript driver
and it turns out we always use the default driver (which is poltergeist
anyway, so actually does understand javascript because it's a "real"
browser).  This means we can also delete the `with_javascript`,
`without_javascript`, and `with_and_without_javascript` helpers because
we don't use them in any tests, and our driver is always a javascript one
anyway.

We do this because our linter suggested that we tidy up the code in
wait_until as it broke the `Lint/HandleExceptions` cop by having an
empty `rescue TimeoutError` block.