Update frontend for updated gems
govuk_navigation_helpers and govuk_publishing_components have changed meaning
that we have to:

- remove the task list AB tests
- amend tests because the schema has changed
- update calls to components which have changed from task list to step by step nav
- rename task list to step_nav or similar in the app

Unfortunately none of these things really work in isolation given the dependency
that govuk_navigation_components and govuk_publishing_components have on each
other for this release.

I've also removed the learn-to-drive-a-car.json file as this was already being
provided by the govuk_navigation_helpers gem.