Set up unicorn config for govuk_app_config
This uses the shared configuration provided by govuk_app_config. I went
out on a limb and decided it was unlikely this app needed to be special
and have the working_directory value set. I couldn't find any evidence
this was required and hopefully if it is we'll find out pretty quick and
can provide the reason in the commit message (which is long lost
currently).