Makes use of usesLicensify API param
For: https://trello.com/c/QswdEVKO/69-2-changes-to-licensing-controller-for-licensify

It was pointed out that some licenses are showing up incorrectly as available in licensify,
with the user being offered the option to start applying for it, although they cannot continue
further than the first step. This is frustrating because the user gets the feeling of being stuck
when in fact they should've never been offered the option in the first place.

The API response already tells us whether the licence should be using Licensify, in a param called
`usesLicensify` so we now check it before showing a start page to the user.

If the param is set to false, we now display a "License unavailable" page instead.