Update frontend to use new step nav helpers
This has moved to govuk_publishing_components as it was closely coupled to
the components.  WE also change our approach a little, as the navigation
information is obtained from the content item's links rather than a hardcoded
file in the navigation helpers gem.

We can remove some tests because we are not responsible for things that the
component does, we just need to ensure that it renders.