Use contextual navigation components
This PR uses the contextual navigation sidebar and breadcrumbs from the
publishing components gem. It removes all the logic for what to display
where and replaces it with calls to the components.

All sidebars now look like the new sidebar. I've removed the extreme
top margin for the sidebar because it looks weird for some pages. We
might want to discuss putting in something smaller.

The sidebar on the /help page is staying the same. It currently has 2
hardcoded links, which can't be set in content-tagger.

https://trello.com/c/R9JMwXC4