Add 'MOT reminder' done page promotion
A new promotion `MOT reminder` needs to be added to the `completed_transaction` done pages. If `mot_reminder` has been selected as the promotion option in `Publisher` the appropriate text will be shown and a `sign up` button, which links to the URL set in the `Promotion choice URL` field in `Publisher`.