Show the local authority URL when available in licensing API
Currently a link to "find your local council" will be shown when the local
authority does not use licensify.  This changed the behaviour to display
a link to the local authority's own website when this is available.

A new value in the API has been added (`usesAuthorityUrl`) which tells
us whether we should display a local authority URL or not.