Ensure B variant of related links shows no related links if suggested related links do not exist.
Previously, a user on the B variant would have seen related links of the control variant if the test variant contained no related links, as `ordered_related_items` would not have been overridden on the content item. This commit fixes that problem, to ensure a fair test.

This commit also renames `RelatedLinksABTest` to `RelatedLinksABTest1` in anticipation of multiple test iterations.

Solo: @karlbaker02