Remove updated date information
"Do a thing" content is our best content and is aimed at the majority of GOV.UK users.
When it's updated though, it is rarely a major update which means the "last updated"
date on the content is often something like 2013.  This does not add anything to
the user's comprehension or confidence that they're looking at something that
is current and relevant.

We're about to publish a "Services" finder which allows users to find this sort of
content more easily, however we won't be adding a date sort to the finder because it's
just not relevant for this sort of content.