Move root route to the top of the file
https://edgeguides.rubyonrails.org/routing.html#using-root states:

> You should put the root route at the top of the file, because it is the
> most popular route and should be matched first.

This is especially pertinent for Frontend because we use routing constraints
that request content items to check the schema of the page to determine
where to send it.

The homepage has its own controller that doesn't need to worry about this
stuff, so let's not slow it down by unnecessarily checking things.

In a similar vein, the component guide is not routed to in production, so
it kind of makes more sense to put it at the bottom of the routes file.