Use unicorn in development
Since the introduction of the puma webserver with govuk_test
(https://github.com/alphagov/govuk_test/pull/1), this app uses puma
instead of unicorn when running `rails server`.

On my machine, a change in the CSS will often result in the server
hanging. My hypothesis is that it's caused by some kind of thread
deadlock caused by the usage of puma. Using unicorn seems to stop the
app from hanging. In any case, we should be using
unicorn because that's what we use in production.