Add QAPage presenter for specific transaction
Adding QAPage formatted information to pages can result in "Q&A" formatted
featured snippets in public search engines.

A [QAPage](https://schema.org/QAPage) has a `mainEntity` that is a [Question](https://schema.org/Question).
The Question may have `suggestedAnswers` and an `acceptedAnswer`. Each of these
uses the [Answer](https://schema.org/Answer) schema.

We're not using suggestedAnswers in this implementation because we're providing
the only answer to our question.

We're not using `upvoteCount` because we haven't got anything that would
allow us to use anything but a made up number. The lack of the `upvoteCount`
does flag as a warning in the testing tools (listed below), but the featured
snippet preview still works OK.

We haven't tried this particular schema before but testing in
https://search.google.com/structured-data/testing-tool/u/0 and
https://search.google.com/test/rich-results indicates that it might work
quite well.

We may expand the use of this in the future (on answers and guides for example),
but this is for [this particular page](https://publisher.publishing.service.gov.uk/editions/5d667a25ed915d23a98aa42b)
at present.