Add FAQ page schema for transaction pages
This is to expose Register to Vote content in a way that can be used more
easily by external consumers.  The content has been recently reworked to
include more things that users are searching for around this topic.

This is only to be activated on /register-to-vote initially, though there's
no reason to not expand this, should research support that it's effective.

The code to extract questions and answers from HTML in the content item
has been lifted from [the FAQPage schema used in guides](https://github.com/alphagov/govuk_publishing_components/blob/a5b0342aa6529c10ce339b9047fb09e7c13f933a/lib/govuk_publishing_components/presenters/machine_readable/faq_page_schema.rb#L45-L76)

It has been tweaked from the guide implementation of the FAQPage schema
in that we're ignoring preamble text in the page (that is, content that
is not preceded by an H2 heading), rather than using a default "question"
for that text.

We're generating the schema here so that we can iterate quickly without
the need to republish the publishing components gem.  We've left the article
schema on transaction pages, and that's OK!