Set the tag on the Sidekiq worker process
At the moment Sidekiq automatically infers the tag using the basename of
the application. In the draft stack this ends up being `draft-frontend`
even though `GOVUK_APP_NAME` remains as `frontend`. This means our
alerting in the draft stack doesn't work because it's looking for a
Sidekiq process called `frontend`, not `draft-frontend`.