Swap FAQ page implementation
We'd originally implemented the FAQPage schema generator to automatically
split the page by section and render each heading as an FAQPage `question`,
and the following text as an `answer`.

Unfortunately, this doesn't quite work for the way we want to use this. Sometimes
the content that provides an `answer` spans a couple of sections.  In this
case, we want to be able to include the content from more than one heading
in the `answer`.

We also want to occasionally omit some text from the FAQ page schema, because
it might be content that helps the page read well, but doesn't add anything
when split into smaller sections.