Add query params and fix content
We're adding a query param to the links in the FAQSchema to help us measure
when traffic is being referred from links parsed from the schema, and when
it's coming from more traditional means.

[The query param (`?foo=bar`) should go before the fragment (`#baz`)](https://tools.ietf.org/html/rfc3986#section-4.2)
where applicable.

We've also removed "your name" from a sentence to match recently updated content.

Before:
> Register to get `your name` on the electoral register. You need to be on the electoral register to vote in elections and referendums.
After:
> Register to get on the electoral register. You need to be on the electoral register to vote in elections and referendums.