Fix conditional reveal for Companies House number
The previous logic showed the "Yes" option selected, even if the user
were yet to make a selection.  Additionally, selecting "No" would show
the word "No" in the number field if the user came back to change their
answers.

This changes the logic so we store both the value of the boolean radio
buttons plus the user's input into the text field.