Increase the timeout for licence applications
Requests to licence_application are timing out in licensify. This means that the licence application forms are not loading
on pages like: `https://www.gov.uk/premises-licence/birmingham`

After some investigation it appears that the request to licensify is taking longer than the default 4 seconds.
This increases the timeout whilst the underlying problem is being investigated.