Remove hardcoded FAQPage schema
We used this approach for the lead up to the election to present better
search results in Google informing people of voting process.  This has a lot
of manual overhead and requires much back and forth between content, product
and tech.  There's little reason for us to do this now that the election has
completed, so let's just revert to the standard schemas.

When we work out a method of better automating this, it should be revisited.