Send EU funding email directly to relevant department
We were sending the departmental email for the EU Funding Form
to the Grants Management Function team at Cabinet Office.  We
now need to send the form directly to the relevant department.

This commit allows us to send emails to specific departments,
based on the grant name selected.  We also need to allow multiple
email addresses to be defined for each grant.