Add AB test for electric cars
This test compares click through rates of people visiting the two done pages:

 - /done/vehicle-tax
 - /done/check-vehicle-tax

We append the variant name on to the end of the URL that leads to the GoUltralow
website so that we they can do some measurement based on what users have
seen here.

Measurement on GOV.UK only occurs if users have opted in to cookies (I can't
speak for what happens on the destination website). People who haven't opted
in will get the control variant which does not add a tracking query parameter
to the link.

The config in en.yml will be changing when we get the final list from DfT.

https://trello.com/c/In6XcrQU/1399-set-up-dft-multi-variant-a-b-test-test-dev-work

Also hooks in with https://trello.com/c/EpuznchJ/1398-dft-for-the-multivariate-test-analytics-tasks