Update travel advice page to show more content
This updates the travel advice index page to modify a few pieces of
static content, and also show the description of the associated content
item in a lead paragraph. Previously the notification div had a custom
margin to align it with the bottom of the search box div. This removes
the padding, since it was proving difficult to maintain.

Related ticket: https://govuk.zendesk.com/agent/tickets/3934135