Add rate limit on specific page
The effect of this code was to put each new visitor to the service start page into one of two buckets based on a random number. 20% of visitors would be put into a bucket where they would see a short message in the place of the service start button, asking them to try again soon. After two minutes, the cookie that identified which bucket the user belongs to would expire, allowing the user to go through the randomised process again, with another 20% chance of being delayed. This amounted to a 4% chance of not being allowed through to the service after 4 minutes.

This solution was chosen as the most pragmatic and low risk solution after some careful consideration of the risks, pros and cons of the alternatives. We understood that users who were unable to run Javascript would not be rate limited, and we determined that this was acceptable.